Fix incorrect proxy ConfigMap name
The mitmproxy ConfigMap implementation previously assumed that
the Deployment and Service name were identical. While this is typically
true in "simple" deployments (such as our testbed, which I will update
at some point to add a test for this), it is not true for complex
deployments.

This is a hotfix, and this fix should be revisited in a future revision
for a more complete fix, including tests and code cleanups.

This fixes #1.